refactor(rte-table): split table geometry into measurement and axis math
`geometry.ts` held two unrelated kinds of code behind one interface: DOM
measurement that needs a live editor, and pure arithmetic over rects that
needs nothing but numbers. Because they shared a module, the pure half
inherited the impure half's imports — `AXIS`, and through it
`@tiptap/pm/tables` — so `resolveDropIndex` was only reachable from the
browser project. Three interacting off-by-one rules verified by three slow
drags, while `coordinates.ts` had a 60-line unit test for four lines of
subtraction.

Split at that seam:

- `internal/measure.ts` — finds tables, reads their rects, converts
  viewport <-> container, and holds the change check that keeps handle
  identity stable across transactions. `coordinates.ts` folds in here;
  `viewportRectToContainerRect` becomes private, as nothing outside used it.
- `internal/axisMath.ts` — the rect model plus every rule computed from it.
  Zero imports: each axis passes an `AxisProjection` describing how it reads
  a rect, so the same arithmetic serves rows and columns.

`AxisSpec` now extends `AxisProjection` and gains `rectsOf`, which retires
`rectsForAxis` and lets a descriptor be handed straight to the arithmetic.
Drops the unused `EMPTY_RECTS` export.

Tests: `axisMath.test.ts` covers boundaries, nearest boundary, drop index,
bounds and span table-driven — including cases nobody would sit through as a
browser drag (dropping onto yourself, dropping above a locked header,
one-row tables, unmeasured slots). 30 assertions in 22ms. The two coordinate
conversions that still have external callers keep their coverage in
`measure.test.ts`; the third retires with the module.

No public surface change — `index.ts` and both call sites are untouched.
All 25 browser tests still pass.
