--- apps/studio/src/features/editing-experience/components/TableDragHandles/internal/__tests__/coordinates.test.ts
@@ -1,60 +1,0 @@
-import {
-  containerRectToViewportRect,
-  viewportPointToContainerPoint,
-  viewportRectToContainerRect,
-} from "../coordinates"
-
-describe("tableEditorGeometry", () => {
-  const containerRect = { top: 100, left: 50 }
-  const scroll = { scrollTop: 40, scrollLeft: 15 }
-
-  it("converts a viewport rect into scroll-aware container coordinates", () => {
-    const viewportRect = {
-      top: 130,
-      left: 90,
-      width: 200,
-      height: 30,
-    }
-
-    const result = viewportRectToContainerRect({
-      rect: viewportRect,
-      containerRect,
-      ...scroll,
-    })
-
-    expect(result).toEqual({
-      top: 70,
-      left: 55,
-      width: 200,
-      height: 30,
-    })
-  })
-
-  it("converts a container rect back into viewport coordinates", () => {
-    const rect = { top: 70, left: 55, width: 200, height: 30 }
-
-    const result = containerRectToViewportRect({
-      rect,
-      containerRect,
-      ...scroll,
-    })
-
-    expect(result).toEqual({
-      top: 130,
-      left: 90,
-      width: 200,
-      height: 30,
-    })
-  })
-
-  it("converts a pointer into the same scroll-aware container coordinates", () => {
-    const result = viewportPointToContainerPoint({
-      clientX: 90,
-      clientY: 130,
-      containerRect,
-      ...scroll,
-    })
-
-    expect(result).toEqual({ x: 55, y: 70 })
-  })
-})
